+++ JasonState/Interfaces/IStateMachine.cs
@@ -1,15 +1,17 @@
-﻿using System.Collections.Generic;
-using JasonState.Models;
 using System;
+using System.Collections.Generic;
+using JasonState.Models;
 
 namespace JasonState.Interfaces
 {
-    public interface IStateMachineBuilder
+    public interface IStateMachine
     {
         IEnumerable<BaseState> BuildMachine(string path);
 
         void AddToContext(Type type);
 
         void AddToContext(IEnumerable<Type> types);
+        
+        void Execute(IEnumerable<BaseState> states);
     }
 }
